Sanitize backend-spi - move auth infrastructure to backend-base/toolkit
-  the call guards, noop providers and stuff used for implementation of
   bear and tiger backends should not be part of SPI
-  the backend-base is the right place for this, established the 'toolkit' part,
   where various infrastructure useful for construction of custom backends
   can be located.
