+++ libs/sdk-backend-tiger/src/backend/workspace/execution/executionFactory.ts
@@ -1,13 +1,9 @@
 // (C) 2019-2020 GoodData Corporation
 
-import {
-    AbstractExecutionFactory,
-    IPreparedExecution,
-    NotImplemented,
-    AuthenticatedCallGuard,
-} from "@gooddata/sdk-backend-spi";
+import { AbstractExecutionFactory, IPreparedExecution, NotImplemented } from "@gooddata/sdk-backend-spi";
 import { IExecutionDefinition, IFilter } from "@gooddata/sdk-model";
 import { TigerPreparedExecution } from "./preparedExecution";
+import { AuthenticatedCallGuard } from "@gooddata/sdk-backend-base";
 
 export class TigerExecution extends AbstractExecutionFactory {
     constructor(private readonly authCall: AuthenticatedCallGuard, workspace: string) {
