+++ libs/sdk-ui-filters/src/DateFilter/utils/OptionUtils.ts
@@ -1,10 +1,11 @@
-// (C) 2019 GoodData Corporation
+// (C) 2019-2020 GoodData Corporation
 import { ExtendedDateFilters } from "../interfaces/ExtendedDateFilters";
 
-export const getDateFilterOptionGranularity = (
+export function getDateFilterOptionGranularity(
     dateFilterOption: ExtendedDateFilters.DateFilterOption,
-): ExtendedDateFilters.DateFilterGranularity =>
-    ExtendedDateFilters.isRelativeDateFilterForm(dateFilterOption) ||
-    ExtendedDateFilters.isRelativeDateFilterPreset(dateFilterOption)
+): ExtendedDateFilters.DateFilterGranularity {
+    return ExtendedDateFilters.isRelativeDateFilterForm(dateFilterOption) ||
+        ExtendedDateFilters.isRelativeDateFilterPreset(dateFilterOption)
         ? dateFilterOption.granularity
         : undefined;
+}
